assistant2: Extract `ActiveThread` view
